requirements.txt
flask==3.0.0
flask-cors==4.0.0
numpy==1.24.3
matplotlib==3.7.1
paho-mqtt==1.6.1
scipy==1.11.1